requirements.txt
fastapi
uvicorn[standard]
pydantic
google-ads
python-dotenv
